+++ gerobug_dashboard/dashboards/apps.py.old
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DashboardsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'dashboards'
